+++ lib/doorkeeper/oauth/authorization_code_request.rb
@@ -63,11 +63,10 @@ module Doorkeeper
       end
 
       def validate_redirect_uri
-        Helpers::URIChecker.native_uri?(redirect_uri) ||
-          Helpers::URIChecker.valid_for_authorization?(
-            redirect_uri,
-            grant.redirect_uri
-          )
+        Helpers::URIChecker.valid_for_authorization?(
+          redirect_uri,
+          grant.redirect_uri
+        )
       end
 
       # if either side (server or client) request pkce, check the verifier