+++ NEWS.md
@@ -19,6 +19,7 @@ User-visible changes worth mentioning.
 - [#1010] Add configuration to enforce configured scopes (`default_scopes` and
   `optional_scopes`) for applications
 - [#1053] Support authorizing with query params in the request `redirect_uri` if explicitly present in app's `Application#redirect_uri`
+- [#1060] Ensure that the native redirect_uri parameter matches with redirect_uri of the client
 
 ## 4.3.1